+++ source/extensions/omni.flux.validator.mass.core/omni/flux/validator/mass/core/executors/__init__.py
@@ -15,5 +15,5 @@
 * limitations under the License.
 """
 
-from .async_executor import AsyncExecutor
-from .process_executor import ProcessExecutor
+from .current_process_executor import CurrentProcessExecutor
+from .external_process_executor import ExternalProcessExecutor